step1/requirements.txt
python-dotenv~=0.19.2
boto3~=1.20.22
botocore~=1.23.22
Unidecode~=1.3.2